Refactor; Initial OOP classes implementation.
