+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,10 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\HenriksHobbyLager",
+    "\\HenriksHobbyLager\\docs",
+    "\\HenriksHobbyLager\\Reports"
+  ],
+  "SelectedNode": "\\HenriksHobbyLager\\Reports\\report.md",
+  "PreviewInSolutionExplorer": false
+}
