+++ .vs/tasks.vs.json
@@ -1,0 +1,15 @@
+{
+  "version": "0.2.1",
+  "tasks": [
+    {
+      "taskLabel": "task-HenriksHobbyLager",
+      "appliesTo": "HenriksHobbyLager/",
+      "type": "launch"
+    },
+    {
+      "taskLabel": "task-HenriksHobbyLager",
+      "appliesTo": "HenriksHobbyLager/",
+      "type": "launch"
+    }
+  ]
+}
