+++ HenriksHobbyLager/HenriksHobbyLager.csproj
@@ -9,9 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="Database\dummy.md" />
-    <Compile Include="Interfaces\dummy.md" />
-    <Compile Include="Models\dummy.md" />
+
   </ItemGroup>
 
 </Project>
