--- HenriksHobbyLager/Models/dummy.md
@@ -1,5 +1,0 @@
-﻿# Dummy
-
-Radera denna fil om du vill.
-
-Den finns bara för att Git ska ta med mappen.
